Get rid of exec by moving the async code to a different module,
Fixes #5.
